Board: remove all boards to keep only one board
