--- game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/board/Connect4Board.java
@@ -1,9 +1,0 @@
-package fr.lecampusnumerique.cda2025.javaalgo.boardgames.board;
-
-public class Connect4Board extends AbstractBoard {
-
-
-    public Connect4Board() {
-        super(6, 7);
-    }
-}
