--- game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/board/GomokuBoard.java
@@ -1,10 +1,0 @@
-package fr.lecampusnumerique.cda2025.javaalgo.boardgames.board;
-
-import fr.lecampusnumerique.cda2025.javaalgo.boardgames.cell.AbstractCell;
-
-public class GomokuBoard extends AbstractBoard {
-
-    public GomokuBoard() {
-        super(15, 15);
-    }
-}
