--- game_board/src/fr/lecampusnumerique/cda2025/javaalgo/boardgames/board/TicTacToeBoard.java
@@ -1,8 +1,0 @@
-package fr.lecampusnumerique.cda2025.javaalgo.boardgames.board;
-
-public class TicTacToeBoard extends AbstractBoard {
-
-    public TicTacToeBoard() {
-        super(3, 3);
-    }
-}
